Adjusted the areaChangePattern
The old pattern was too far down the function it was in.  Any changes further up in the function had the potential to move it from its 4 byte offset and prevent the scan from finding it.  The new offset is much closer to the start of the function so it should be safer.  Also, it works in 2.5.0.