handle exception for items on the ground
List of ground items that are in UIRoot can get In-valid (null) from valid very fast if you hover mouse on them quickly.
Due to this GetToolTipType() throws an exception because it tries to read from an invalid object.
This try catch will return None if there is an exception i.e. object (or object UI) doesn't exists anymore.