A new way of calculating mod tier
In this method we isolate related mods by following properties

0: Get all mods from the category of mods.
Let's call them mods under consideration.

1: After removing the symbols, numbers from the mod name
of the item and other mod's name under consideration,
they must match.

2: Using mod tags we see if mods under consideration is allowed
on the item or not. This is a 4 step process
i: We see mods under consideration tagchance for baseitemtype must be
either > 0 or tag not found.
ii: We see mods under consideration tagchance for item tags must be
either > 0 or tag not found.
iii: We see mods under consideration tagchace for item category must be
either > 0 or tag not found.
iV: We see that default mod tag must be greater than 0.
Note: if we find any tagchance == 0. we ignore that mod.

In the end we calculate totalMods and mod Tier with total mods left in the list.