Fixes to failed ReadProcessMemory calls
Memory.cs - Added breaks (and thus return 0) if the result of the last read was 0. Reading address 0 + offset doesn't really work. This mostly seemed to be occuring on Entity.IsValid (specifically the internalEntity.IsValid call). The address sent in had a valid value, but reading the long returned 0 and it attempted to read [0+32}+0.

ReadMemoryObject.cs - Objects with the address of 0 will be equal to null. There may be a better way to do this, but there are many cases where calls such as Game.IngameState.IngameUi.ReadObjectAt<ItemsOnGroundLabelElement>(0xD58); are made and the return checked for null. As far as I can tell, the ObjectAt methods will never return null.